Move terminating logic into invokable class
